+++ packages/gov-ui-kit/src/core/components/index.ts
@@ -1,5 +1,6 @@
 export * from './alerts';
 export * from './avatars';
+export * from './breadcrumbs';
 export * from './button';
 export * from './cards';
 export * from './checkbox';